@@ -9,11 +9,11 @@
     "type": "library",
     "config": {
         "platform": {
-            "php": "8.0.0"
+            "php": "8.1.0"
         }
     },
     "require": {
-        "php": ">=8.0"
+        "php": ">=8.1"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^2.0",
